backend/app/main.py
"""FastAPI application for the Agentic Form Filler."""

import os
from pathlib import Path
from contextlib import asynccontextmanager

from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from dotenv import load_dotenv

# Load environment variables
load_dotenv()


@asynccontextmanager
async def lifespan(app: FastAPI):
    """Application lifespan handler."""
    # Startup: ensure upload directory exists
    Path("uploads").mkdir(exist_ok=True)
    yield
    # Shutdown: cleanup could go here


app = FastAPI(
    title="Agentic Form Filler",
    description="AI-powered movie order form filler using Google ADK",
    version="1.0.0",
    lifespan=lifespan
)

# Configure CORS for frontend
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:5173", "http://127.0.0.1:5173"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


@app.get("/health")
async def health_check():
    """Health check endpoint."""
    return {"status": "healthy"}


# Import and include routers after app is created to avoid circular imports
from app.routes import router
app.include_router(router, prefix="/api")


