backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-docx==1.1.2
python-multipart==0.0.19
pycountry==24.6.1
google-adk==0.3.0
python-dotenv==1.0.1
docx-mailmerge2==0.8.0

